Moved Environment interface to common
