+++ common/src/commonMain/kotlin/com/thebrownfoxx/neon/common/environment/Environment.kt
@@ -1,0 +1,8 @@
+package com.thebrownfoxx.neon.common.environment
+
+interface Environment {
+    operator fun get(key: EnvironmentKey): String
+}
+
+@JvmInline
+value class EnvironmentKey(val name: String)
