--- server/application/src/main/kotlin/com/thebrownfoxx/neon/server/application/environment/Environment.kt
@@ -1,5 +1,0 @@
-package com.thebrownfoxx.neon.server.application.environment
-
-interface Environment {
-    operator fun get(key: EnvironmentKey): String
-}
